requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.3
python-dotenv==1.0.0
Werkzeug==2.2.3
email_validator==2.0.0
PyJWT==2.6.0